confirm delete added with sweetalerts

+++ client/src/components/List.vue
@@ -14,41 +14,47 @@
 </template>
 
 <script>
-export default {
-  name: "List",
-  computed: {
-    lists() {
-      let data = this.$store.state.lists;
-      return data;
-    },
-    activeBoard() {
-      return this.$store.state.activeBoard;
-    }
-  },
+  import NotificationService from "../NotificationService"
 
-  props: ["listData", "listIndex"],
-  methods: {
-    deleteList() {
-      this.$store.dispatch("deleteList", this.listData);
+  export default {
+    name: "List",
+    computed: {
+      lists() {
+        let data = this.$store.state.lists;
+        return data;
+      },
+      activeBoard() {
+        return this.$store.state.activeBoard;
+      }
+    },
+
+    props: ["listData", "listIndex"],
+    methods: {
+      async deleteList() {
+        if (await NotificationService.confirmDelete()) {
+          this.$store.dispatch("deleteList", this.listData);
+        }
+      }
     }
-  }
-};
+  };
 </script>
 
 <style>
-.list-card {
-  width: 35vh;
-  height: 54vh;
-  margin-bottom: 0vh;
-  background-image: URL("../assets/aspenbark.jpg");
-  box-shadow: 3px 3px 7px 2px black;
-}
-.list-font {
-  font-family: "Boogaloo", cursive;
-  color: rgb(77, 51, 19);
-  text-shadow: 2px 2px rgb(255, 255, 255);
-}
-.bottom {
-  bottom: 0;
-}
+  .list-card {
+    width: 35vh;
+    height: 54vh;
+    margin-bottom: 0vh;
+    background-image: URL("../assets/aspenbark.jpg");
+    box-shadow: 3px 3px 7px 2px black;
+  }
+
+  .list-font {
+    font-family: "Boogaloo", cursive;
+    color: rgb(77, 51, 19);
+    text-shadow: 2px 2px rgb(255, 255, 255);
+  }
+
+  .bottom {
+    bottom: 0;
+  }
 </style>
